Refactor payment response amount to double and update payment link creation logic

+++ src/main/java/com/lld/payment_service/dtos/responsedtos/PaymentResponseDto.java
@@ -1,5 +1,6 @@
 package com.lld.payment_service.dtos.responsedtos;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,5 +10,5 @@ public class PaymentResponseDto {
 
     private String url;
     private String status;
-    private long amount;
+    private double amount;
 }
